CleanUp and renaming appropriately.
