+++ CsCore/PlainNetClassLib/PlainNetClassLib.csproj
@@ -39,7 +39,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
-	<None Include="..\..\README.md" Pack="true" PackagePath="\"/>
+	<None Include="..\..\README.md" Pack="true" PackagePath="\" />
   </ItemGroup>
